build.sh: implement runtime vars save/load, rework makepkg+ccache.conf
Implement a generalized mechanism for passing variables from the
top-level build.sh process to subroutine processes instead of using
argument passthrough and hoping that the subroutines come up with
the same final values for all relevant arguments.

For now, use this mechanism to clean up handling of makepkg+ccache.conf
instead of exporting a magic environment variable and picking it up.