build.sh: compute pacman db path and pass it to aur-repo
Compute the path to the /var/lib/pacman/sync that's effectively used
in the package build environment (i.e. taking chroot path and custom
pacman.conf into account), and use it to set `aur repo --root`.

This way, aur-repo can be used to query both Arch repositories and the
current target repository, regardless of whether the host pacman.conf
has the target repository added (which is not necessarily the case).