build.sh: fix chroot initial setup
Our `aur chroot ... -- -uu` hack to permit package downgrades when
updating the chroot base breaks when the chroot needs to be created
because in this case the positional arguments, when present, are treated
as the list of packages to install instead of the default groups.

Fix this by (ab)using `aur chroot --path` to determine whether the
chroot needs to be created or updated and only performing the
`aur-chroot -- -uu` hack in the update path.