build.sh: autobump: rewrite again, now matching on pkgnames
The previous implementation failed to consider the fact that a pkgname
may "change hands" and be provided by a different pkgbase than in the
official repos (or in the previous iteration of the custom repo, for
that matter). Handle this and extract pkgvers of all matching pkgnames
rather than of the matching pkgbase (accounting for split packages).

While at it, unify handling of the custom repo and the official repos.